Finally fix doc error
Avoid re-installing things that are already on the runner (i.e. python & pip, build-essential, clang) and use the versions that are pre-configured.  Printing `compiler.get_include_args()` in conf.py showed that clang 12 was trying to use clang-11 libs, so switched to clang-11 & python3-clang 11 to finally solve the problem.  This commit just removes the logging.